requirements.txt
absl-py==1.1.0
astor==0.8.1
backcall==0.2.0
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cycler==0.11.0
decorator==5.1.1
Flask==2.1.2
Flask-Cors==3.0.10
fonttools==4.33.3
gast==0.2.2
gcloud==0.18.3
gevent==21.12.0
google-pasta==0.2.0
googleapis-common-protos==1.56.3
greenlet==1.1.2
grpcio==1.47.0
gunicorn==20.1.0
h5py==2.10.0
httplib2==0.20.4
idna==3.3
imageio==2.19.3
importlib-metadata==4.12.0
ipython==7.34.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jws==0.1.3
Keras==2.1.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
networkx==2.6.3
numpy==1.21.6
oauth2client==4.1.3
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
Pygments==2.12.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dateutil==2.8.2
python-jwt==2.0.1
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.8
scikit-image==0.16.2
scipy==1.7.3
six==1.16.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
tifffile==2021.11.2
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.10
wcwidth==0.2.5
Werkzeug==2.1.2
wincertstore==0.2
wrapt==1.14.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0